+++ client/pico_oxr/app/src/main/cpp/streaming/streaming_client.h
@@ -38,8 +38,8 @@ struct streaming_client
 	std::unique_ptr<wivrn_session_pico> session;
 
 	pico_blit_pipeline blit_pipeline;
-	int eye_width = 2048;
-	int eye_height = 2160;
+	std::atomic<int> eye_width{2048};
+	std::atomic<int> eye_height{2160};
 	std::atomic<int> stream_eye_width{2048};
 	std::atomic<int> stream_eye_height{2160};
 	std::atomic<bool> streaming{false};